requirement.txt
Flask
flask_session
mysql-connector-python
werkzeug
pyotp
qrcode
reportlab
